+++ src/main/resources/synthea.properties
@@ -10,28 +10,28 @@ exporter.years_of_history = 10
 exporter.split_records = false
 exporter.split_records.duplicate_data = false
 exporter.ccda.export = false
-exporter.fhir.export = true
+exporter.fhir.export = false
 exporter.fhir_stu3.export = false
 exporter.fhir_dstu2.export = false
 exporter.fhir.use_shr_extensions = false
 exporter.fhir.use_us_core_ig = false
-exporter.fhir.transaction_bundle = true
+exporter.fhir.transaction_bundle = false
 exporter.fhir.bulk_data = false
-exporter.groups.fhir.export = false
-exporter.hospital.fhir.export = true
+exporter.hospital.fhir.export = false
 exporter.hospital.fhir_stu3.export = false
 exporter.hospital.fhir_dstu2.export = false
-exporter.practitioner.fhir.export = true
+exporter.practitioner.fhir.export = false
 exporter.practitioner.fhir_stu3.export = false
 exporter.practitioner.fhir_dstu2.export = false
-exporter.csv.export = false
+exporter.csv.export = true
+exporter.cpcds.export = false
+exporter.cpcds.postprocess = false
+exporter.cpcds.append_mode = false
+exporter.cpcds.folder_per_run = false
 # if exporter.csv.append_mode = true, then each run will add new data to any existing CSVs. if false, each run will clear out the files and start fresh 
 exporter.csv.append_mode = false
 # if exporter.csv.folder_per_run = true, then each run will have CSVs placed into a unique subfolder. if false, each run will only use the top-level csv folder
 exporter.csv.folder_per_run = false
-exporter.cpcds.export = false
-exporter.cpcds.append_mode = false
-exporter.cpcds.folder_per_run = false
 exporter.cdw.export = false
 exporter.text.export = false
 exporter.text.per_encounter_export = false
@@ -80,7 +80,7 @@ generate.only_dead_patients = false
 generate.track_detailed_transition_metrics = false
 
 # If true, person names have numbers appended to them to make them more obviously fake
-generate.append_numbers_to_person_names = true
+generate.append_numbers_to_person_names = false
 
 # if true, the entire population will use veteran prevalence data
 generate.veteran_population_override = false
@@ -156,8 +156,8 @@ generate.providers.primarycare.default_file = providers/primary_care_facilities.
 generate.providers.selection_behavior = nearest
 
 # maximum distance to look for a provider for a given patient, in km
-# set to 10 degrees lat/lon to support the model that veterans only seek care at VA facilities
-generate.providers.maximum_search_distance = 32
+# set to 2000km to support the model that veterans only seek care at VA facilities
+generate.providers.maximum_search_distance = 2000
 
 # Payers
 generate.payers.insurance_companies.default_file = payers/insurance_companies.csv
@@ -184,6 +184,3 @@ lifecycle.adherence.baseline = 0.05
 # set this to true to enable randomized "death by natural causes"
 # highly recommended if "only_dead_patients" is true
 lifecycle.death_by_natural_causes = false
-
-# Use physiology simulations to generate some VitalSigns
-physiology.generators.enabled = false
